[LDAP] ExtLdap pagination support
Support pagination in the ExtLdap adapter. Pagination is requested by setting
the 'pageSize' option on the Query.

The pageSize and maxItems options will cooperate as expected. If the pageSize
exceeds maxItems, pagination controls will not be sent to the server. The
total number of results will be capped by maxItems, whether all pages for
the query are retrieved or not.